Refactor: js, jsx파일 분리
