+++ .eslintrc.json
@@ -46,5 +46,12 @@
         "ignoreTemplateLiterals": true
       }
     ]
+  },
+  "settings": {
+    "import/resolver": {
+      "node": {
+        "extensions": [".js", ".jsx"]
+      }
+    }
   }
 }
